requirements.txt
Flask-Migrate
Flask-RESTful
Flask-SQLAlchemy
Flask-Script
Flask
Jinja2
SQLAlchemy
Werkzeug
alembic
flasgger
itsdangerous
psycopg2
pytz
requests
python-dotenv
setproctitle
six

autopep8
isort
pytest
pip-upgrader
pytest-cov
pytest-sugar
safety